remove unneeded pages test. make browser test in the expected format
