+++ test/components-browser/fixtures/force-update-dom-exception-3/components/Box-even.marko
