--- test/components-pages/fixtures/force-update-dom-exception-3/components/app/components/Box-even.marko
@@ -1,1 +1,0 @@
-<span style={color:'grey'}>&#9608;</span>
