--- test/components-pages/fixtures/force-update-dom-exception-3/components/app/components/Box-odd.marko
@@ -1,1 +1,0 @@
-<span style={color:'silver'}>&#9608;</span>
