--- test/components-pages/fixtures/force-update-dom-exception-3/components/app/index.marko
@@ -1,20 +1,0 @@
-class {
-  onCreate() {
-    this.array = [0]
-  }
-  onMount() {
-    window.appComponent = this
-  }
-}
-<div.ArrayModel>
-  ${component.array.length}
-  <hr>
-  <for(value in component.array)>
-      ${value}
-      <for(let i = 0 ; i < value; i++)>
-        <Box-even if( i % 2 == 0)/>
-        <Box-odd else/>
-      </for>
-      <hr>
-  </for>
-</div>
