--- test/components-pages/fixtures/force-update-dom-exception-3/template.marko
@@ -1,10 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-      <meta charset="UTF-8">
-  </head>
-  <body>
-    <app/>
-    <init-components immediate/>
-  </body>
-</html>
